pyproject.toml
[tool.poetry]
name = "finalproject"
version = "0.1.0"
description = ""
authors = ["SamuelAlvianM <oensam07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "8.3.0"
bcrypt = "4.1.2"
python-dotenv = "1.0.1"
cerberus = "1.3.5"
waitress = "^3.0.0"
flask = "3.0.2"
blinker = "1.7.0"
cachetools = "5.3.2"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
docopt = "0.6.2"
flask-jwt-extended = "4.6.0"
flask-login = "0.6.3"
flask-sqlalchemy = "3.1.1"
markupsafe = "2.1.5"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pyjwt = "2.8.0"
requests = "2.31.0"
sqlalchemy = "2.0.25"
typing-extensions = "4.9.0"
urllib3 = "2.2.0"
werkzeug = "3.0.1"
yarg = "0.1.9"
flask-migrate = "^4.0.7"
flask-cors = "^4.0.0"
pytest-mock = "^3.14.0"
pydantic = "^2.7.1"
pytest-flask = "^1.3.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
